cleanup files and refactor
