+++ src/app/article/[slug]/page.tsx
@@ -4,18 +4,19 @@ import React from 'react';
 import { useParams } from 'next/navigation';
 import { Header } from '@/components/Header';
 import { RevisionList } from '@/components/RevisionList';
-import { SnapshotLink } from '@/components/SnapshotLink';
 import { useArticleRevisions } from '@/hooks/useArticleRevisions';
-import { Revision } from '@/types/revisions';
+import { ArticleRevision } from '@/types/articleRevisions';
+import { useRouter } from 'next/navigation';
 
 const ArticlePage = () => {
   const params = useParams();
   const slug = params?.slug as string;
   const title = decodeURIComponent(slug.replace(/_/g, ' '));
   const { revisions, isLoading } = useArticleRevisions(title);
+  const router = useRouter();
 
-  const handleRevisionSelect = (revision: Revision) => {
-    console.log('Selected revision:', revision);
+  const handleRevisionSelect = (revision: ArticleRevision) => {
+    router.push(`/article/${revision.titleSlug}/snapshots/${revision.id}`);
   };
 
   return (
@@ -23,12 +24,6 @@ const ArticlePage = () => {
       <Header />
       <div className="container mx-auto px-4 py-8">
         <h1 className="text-3xl font-bold mb-6">Article Revisions: {title}</h1>
-        {revisions.length > 0 && (
-          <SnapshotLink 
-            articleId={revisions[0].id.toString()} 
-            title={title}
-          />
-        )}
         <div className="mt-8">
           {isLoading ? (
             <div>Loading revisions...</div>
