+++ src/types/articleRevisions.ts
@@ -1,0 +1,9 @@
+export interface ArticleRevision {
+  id: number;
+  title: string;
+  titleSlug: string;
+  parentId?: number;
+  user: string;
+  timestamp: string;
+  comment: string;
+}
